+++ app/ThirdParty/WooCommerce.php
@@ -12,6 +12,7 @@ defined( 'ABSPATH' ) || exit;
 
 use ExternalFilesInMediaLibrary\ExternalFiles\Files;
 use ExternalFilesInMediaLibrary\Plugin\Helper;
+use ExternalFilesInMediaLibrary\Plugin\Log;
 use ExternalFilesInMediaLibrary\Plugin\Settings\Settings;
 use WC_Product;
 
@@ -84,11 +85,18 @@ class WooCommerce {
 	 * @return void
 	 */
 	public function init_woocommerce(): void {
+		// get settings object.
 		$settings_obj             = Settings::get_instance();
+
+		// add settings tab for WooCommerce.
 		$woocommerce_settings_tab = $settings_obj->add_tab( 'woocommerce' );
 		$woocommerce_settings_tab->set_title( __( 'WooCommerce', 'external-files-in-media-library' ) );
+
+		// add section for WooCommerce settings.
 		$woocommerce_settings_section = $woocommerce_settings_tab->add_section( 'eml_woocommerce_settings' );
 		$woocommerce_settings_section->set_title( __( 'WooCommerce', 'external-files-in-media-library' ) );
+
+		// add setting to enable the WooCommerce-support.
 		$woocommerce_settings_setting = $settings_obj->add_setting( 'eml_woocommerce' );
 		$woocommerce_settings_setting->set_section( $woocommerce_settings_section );
 		$woocommerce_settings_setting->set_type( 'integer' );
@@ -115,18 +123,27 @@ class WooCommerce {
 			return $data;
 		}
 
+		// log event.
+		Log::get_instance()->create( __( 'Trying to import main image of WooCommerce product as external image.', 'external-files-in-media-library' ), $data['raw_image_id'], 'info', 2 );
+
+		// get the files object.
+		$external_files_obj = Files::get_instance();
+
 		// add the image and bail if it was not successfully.
-		if ( ! Files::get_instance()->add_url( $data['raw_image_id'] ) ) {
+		if ( ! $external_files_obj->add_url( $data['raw_image_id'] ) ) {
 			return $data;
 		}
 
 		// get the external file object for the given main image.
-		$external_file_obj = Files::get_instance()->get_file_by_url( $data['raw_image_id'] );
+		$external_file_obj = $external_files_obj->get_file_by_url( $data['raw_image_id'] );
 
 		// bail if file for the URL does not exist.
 		if ( ! $external_file_obj ) {
 			return $data;
 		}
+
+		// log event.
+		Log::get_instance()->create( __( 'WooCommerce Product main image has been imported as external file.', 'external-files-in-media-library' ), $data['raw_image_id'], 'info', 2 );
 
 		// remove raw_image_id to prevent usage through WooCommerce.
 		unset( $data['raw_image_id'] );
@@ -152,6 +169,9 @@ class WooCommerce {
 			return;
 		}
 
+		// log event.
+		Log::get_instance()->create( __( 'Save external image file as WooCommerce Product main image.', 'external-files-in-media-library' ), '', 'info', 2 );
+
 		// set the given main image on product.
 		wc_product_attach_featured_image( $data['eml_file'], $product );
 	}
@@ -173,6 +193,9 @@ class WooCommerce {
 		if ( ! is_array( $data['raw_gallery_image_ids'] ) ) {
 			return $data;
 		}
+
+		// log event.
+		Log::get_instance()->create( __( 'Trying to save WooCommerce product gallery images as external files.', 'external-files-in-media-library' ), '', 'info', 2 );
 
 		// collect all images.
 		$gallery_images = array();
@@ -204,6 +227,9 @@ class WooCommerce {
 			return $data;
 		}
 
+		// log event.
+		Log::get_instance()->create( __( 'WooCommerce product gallery images has been saved as external files.', 'external-files-in-media-library' ), '', 'info', 2 );
+
 		// add custom column which we use later to assign the media files to the product.
 		$data['eml_files'] = $gallery_images;
 
@@ -220,10 +246,15 @@ class WooCommerce {
 	 * @return void
 	 */
 	public function add_product_gallery_images( WC_Product $product, array $data ): void {
+		// bail if no gallery images are given.
 		if ( empty( $data['eml_files'] ) ) {
 			return;
 		}
 
+		// log event.
+		Log::get_instance()->create( __( 'WooCommerce product gallery images will be saved as external files on the product.', 'external-files-in-media-library' ), '', 'info', 2 );
+
+		// add them to the product as gallery.
 		$product->set_gallery_image_ids( $data['eml_files'] );
 		$product->save();
 	}